'cookiesFirstPart'

+++ ecommerce/store/models.py
@@ -11,7 +11,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True) # If a product is digital, we dont neet to put it on the shipping
     image = models.ImageField(null=True, blank=True)
 
@@ -69,7 +69,7 @@ class OrderItem(models.Model):
     date_added = models.DateField(auto_now_add=True) # Select the current date
 
     @property
-    def get_total(self): # Show the current price of the arrow
+    def get_total(self): # Show the current price of the row
         total = self.product.price * self.quantity
         return total
